♻️ Split multiple-choice presentational component
